requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1
Pillow>=10.0.0
requests>=2.31.0
python-decouple>=3.8
django-extensions>=3.2.0
celery>=5.3.0
redis>=4.6.0
django-ratelimit>=4.0.0
pyotp>=2.8.0
qrcode>=7.4.2
